CAN Transceiver Flow (#88)
* Rework CAN frame parsing

* Better printing

* Restructure inheritance

* Unit tests and utils

* More unit tests

* Clean up can_frame_parser

* Fix errors

* Move utils

* Test CanbusTransceiver with tmp file

* Restructure CAN Sim Interfacing

* Cleanup const and references

* Connect CanTransceiver to ros

* Rename namespace CAN to CAN_FP

* Added CAN Transceiver to ROS network

* Ready for review

* Minor typos

* Address PR comments

+++ projects/remote_transceiver/test/test_remote_transceiver.cpp
@@ -627,7 +627,7 @@ TEST_F(TestHTTP, TestPostSensorsMult)
     std::array<SailbotDB::RcvdMsgInfo, NUM_REQS> expected_info;
 
     // Prepare all queries
-    for (size_t i = 0; i < NUM_REQS; i++) {
+    for (int i = 0; i < NUM_REQS; i++) {
         auto [rand_sensors, rand_info] = genRandData();
         expected_sensors[i]            = rand_sensors;
         expected_info[i]               = rand_info;
@@ -648,7 +648,7 @@ TEST_F(TestHTTP, TestPostSensorsMult)
     }
 
     // Send all requests at once
-    for (size_t i = 0; i < NUM_REQS; i++) {
+    for (int i = 0; i < NUM_REQS; i++) {
         req_threads[i] = std::thread([&queries, &res_statuses, i]() {
             std::string query = queries[i];
             res_statuses[i]   = http_client::post(
